Refactor MSN to a plugin form
