+++ StreamCompanionTypes/Interfaces/IOsuEventSource.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using StreamCompanionTypes.DataTypes;
+
+namespace StreamCompanionTypes.Interfaces
+{
+    public interface IOsuEventSource
+    {
+        EventHandler<MapSearchArgs> NewOsuEvent { get; set; }
+    }
+}
