+++ osu!StreamCompanion/Code/Core/Maps/Processing/MapStringFormatter.cs
@@ -10,21 +10,34 @@ using StreamCompanionTypes.Interfaces;
 namespace osu_StreamCompanion.Code.Core.Maps.Processing
 {
     //TODO: refactor to use list of IOsuEventSource providers instead.
-    public class MapStringFormatter : IModule, IMsnGetter, ISettings,IDisposable
+    public class MapStringFormatter : IModule, ISettings,IDisposable
     {
         private readonly SettingNames _names = SettingNames.Instance;
         private ILogger _logger;
         private readonly MainMapDataGetter _mainMapDataGetter;
+        private readonly List<IOsuEventSource> _osuEventSources;
         private ISettingsHandler _settings;
         private string _lastMsnString = "";
         private Thread ConsumerThread;
-        private ConcurrentStack<MapSearchArgs> TasksMsn = new ConcurrentStack<MapSearchArgs>();
+        private ConcurrentStack<MapSearchArgs> Tasks = new ConcurrentStack<MapSearchArgs>();
 
-        public MapStringFormatter(MainMapDataGetter mainMapDataGetter)
+        public MapStringFormatter(MainMapDataGetter mainMapDataGetter,List<IOsuEventSource> osuEventSources)
         {
             _mainMapDataGetter = mainMapDataGetter;
+            _osuEventSources = osuEventSources;
+            foreach (var source in osuEventSources)
+            {
+                source.NewOsuEvent+=NewOsuEvent;
+            }
             ConsumerThread = new Thread(ConsumerTask);
         }
+
+        private void NewOsuEvent(object sender, MapSearchArgs mapSearchArgs)
+        {
+            Tasks.Clear();
+            Tasks.Push(mapSearchArgs);
+        }
+
         public bool Started { get; set; }
         public void Start(ILogger logger)
         {
@@ -36,59 +49,7 @@ namespace osu_StreamCompanion.Code.Core.Maps.Processing
             _settings = settings;
         }
 
-        public void SetNewMsnString(Dictionary<string, string> osuStatus)
-        {
-            /*osuStatus["artist"]
-            osuStatus["title"]
-            osuStatus["diff"]
-            osuStatus["status"]*/
-
-            OsuStatus status = osuStatus["status"] == "Listening" ? OsuStatus.Listening
-                : osuStatus["status"] == "Playing" ? OsuStatus.Playing
-                : osuStatus["status"] == "Watching" ? OsuStatus.Watching
-                : osuStatus["status"] == "Editing" ? OsuStatus.Editing
-                : OsuStatus.Null;
-
-            osuStatus["raw"] = string.Format("{0} - {1}", osuStatus["title"], osuStatus["artist"]);
-            bool isFalsePlay;
-            lock (this)
-            {
-                isFalsePlay = IsFalsePlay(osuStatus["raw"], status, _lastMsnString);
-            }
-            if (isFalsePlay)
-            {
-                _logger.Log(">ignoring second MSN string...", LogLevel.Advanced);
-            }
-            else
-            {
-                _lastMsnString = osuStatus["raw"];
-                _logger.Log("", LogLevel.Advanced);
-                string result = ">Got ";
-                foreach (var v in osuStatus)
-                {
-                    if (v.Key != "raw") result = result + $"{v.Key}: \"{v.Value}\" ";
-                }
-                _logger.Log(result, LogLevel.Basic);
-                while (_settings.Get<bool>(_names.LoadingRawBeatmaps))
-                {
-                    Thread.Sleep(200);
-                }
-                var searchArgs = new MapSearchArgs()
-                {
-                    Artist = osuStatus["artist"] ?? "",
-                    Title = osuStatus["title"] ?? "",
-                    Diff = osuStatus["diff"] ?? "",
-                    Raw = osuStatus["raw"] ?? "",
-                    Status = status,
-
-                };
-
-                //Throw away any unprocessed tasks- I'm only intrested in processing current one.
-                TasksMsn.Clear();
-                TasksMsn.Push(searchArgs);
-
-            }
-        }
+        
         public void ConsumerTask()
         {
             try
@@ -97,7 +58,7 @@ namespace osu_StreamCompanion.Code.Core.Maps.Processing
                 MapSearchResult searchResult;
                 while (true)
                 {
-                    if (TasksMsn.TryPop(out searchArgs))
+                    if (Tasks.TryPop(out searchArgs))
                     {
                         searchResult = _mainMapDataGetter.FindMapData(searchArgs);
                         _mainMapDataGetter.ProcessMapResult(searchResult);
@@ -113,45 +74,7 @@ namespace osu_StreamCompanion.Code.Core.Maps.Processing
             {
             }
         }
-        #region MSN double-send fix
-        public class MapArgs
-        {
-            public string MapName;
-            public OsuStatus MapAction;
-        }
-        //osu! MSN double-send detection
-        private readonly string[] _lastListened = new string[2];
-        bool IsFalsePlay(string msnString, OsuStatus msnStatus, string lastMapString)
-        {
-            lock (_lastListened)
-            {
-                // if we're listening to a song AND it's not already in the first place of our Queue
-                if (msnStatus == OsuStatus.Listening && msnString != _lastListened[0])
-                {
-                    //first process our last listened song "Queue" 
-                    _lastListened[1] = _lastListened[0];
-                    _lastListened[0] = msnString;
-                }
-                //we have to be playing for bug to occour...
-                if (msnStatus != OsuStatus.Playing)
-                    return false;
-                //if same string is sent 2 times in a row
-                if (msnString == lastMapString)
-                {
-                    //this is where it gets checked for actual bug- Map gets duplicated only when we just switched from another song
-                    //so check if we switched by checking if last listened song has changed 
-                    if (_lastListened[0] != _lastListened[1])
-                    {
-                        //to avoid marking another plays(Retrys) as False- we "break" our Queue until we change song.
-                        _lastListened[1] = _lastListened[0];
-                        return true;
-                    }
-                }
-                return false;
-            }
-        }
-
-        #endregion //MSN FIX
+        
 
         public void Dispose()
         {
