+++ osu!StreamCompanion/osu!StreamCompanion.csproj
@@ -200,7 +200,6 @@
     <Compile Include="Code\Modules\FirstRun\FirstRun.cs" />
     <Compile Include="Code\Modules\MapDataReplacements\Map\MapReplacement.cs" />
     <Compile Include="Code\Core\Maps\Processing\MapStringFormatter.cs" />
-    <Compile Include="Code\Core\Maps\MSN.cs" />
     <Compile Include="Code\Modules\MapDataFinders\SqliteData\SqliteDataFinder.cs" />
     <Compile Include="Code\Modules\MapDataFinders\osuMemoryID\MemoryDataFinderSettings.cs">
       <SubType>UserControl</SubType>
